+++ src/pages/Search/ui/SearchPage.style.ts
